--- submarine-server/server-core/src/main/resources/mybatis-config-metastore.xml
@@ -1,59 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE configuration PUBLIC '-//mybatis.org//DTD Config 3.0//EN'
-    'http://mybatis.org/dtd/mybatis-3-config.dtd'>
-<configuration>
-  <settings>
-    <setting name="cacheEnabled" value="true"/>
-    <setting name="lazyLoadingEnabled" value="false"/>
-    <setting name="aggressiveLazyLoading" value="true"/>
-    <setting name="logImpl" value="STDOUT_LOGGING"/>
-  </settings>
-
-  <typeAliases>
-    <package name="com.github.pagehelper.model"/>
-  </typeAliases>
-
-  <plugins>
-    <plugin interceptor="com.github.pagehelper.PageInterceptor">
-      <property name="helperDialect" value="mysql"/>
-      <property name="offsetAsPageNum" value="true"/>
-      <property name="rowBoundsWithCount" value="true"/>
-    </plugin>
-  </plugins>
-
-  <environments default="development">
-    <environment id="development">
-      <transactionManager type="JDBC"/>
-      <dataSource type="POOLED">
-        <property name="driver" value="${jdbc.driverClassName}"/>
-        <property name="url" value="${jdbc.url}"/>
-        <property name="username" value="${jdbc.username}"/>
-        <property name="password" value="${jdbc.password}"/>
-        <property name="poolPingQuery" value="SELECT NOW()"/>
-        <property name="poolPingEnabled" value="true"/>
-      </dataSource>
-    </environment>
-  </environments>
-
-  <mappers>
-    <mapper resource='org/apache/submarine/database/mappers/EnvironmentMapper.xml'/>
-  </mappers>
-</configuration>
