--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/EnvironmentMapper.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.environment.mappers.EnvironmentMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-  	<id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="resultMap" type="org.apache.submarine.server.environment.database.entity.EnvironmentEntity" extends="BaseEntityResultMap">
-    <result column="environment_name" jdbcType="VARCHAR" property="environmentName" />
-    <result column="environment_spec" jdbcType="VARCHAR" property="environmentSpec" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, environment_name, environment_spec, create_by, create_time, update_by, update_time
-  </sql>
-
-  <select id="selectAll" parameterType="java.lang.String" resultMap="resultMap">
-    select
-    <include refid="Base_Column_List" />
-    from environment
-  </select>
-
-  <select id="select" parameterType="java.lang.String" resultMap="resultMap">
-    select
-    <include refid="Base_Column_List" />
-    from environment
-    where environment_name = #{environment_name,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="delete" parameterType="java.lang.String">
-    delete from environment
-    where environment_name = #{environment_name,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.environment.database.entity.EnvironmentEntity">
-    insert into environment (id, environment_name, environment_spec,
-      create_by, create_time, update_by, update_time)
-    values (#{id,jdbcType=VARCHAR}, #{environmentName,jdbcType=VARCHAR}, #{environmentSpec,jdbcType=VARCHAR},
-     #{createBy,jdbcType=VARCHAR}, now(), #{updateBy,jdbcType=VARCHAR}, now())
-  </insert>
-
-  <update id="update" parameterType="org.apache.submarine.server.environment.database.entity.EnvironmentEntity">
-    update environment
-    <set>
-      <if test="environmentSpec != null">
-        environment_spec = #{environmentSpec,jdbcType=VARCHAR},
-      </if>
-      update_time = now()
-    </set>
-    where environment_name = #{environmentName,jdbcType=VARCHAR}
-  </update>
-
-</mapper>
