--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/ExperimentMapper.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.experiment.mappers.ExperimentMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="ExperimentEntityResultMap" type="org.apache.submarine.server.experiment.database.entity.ExperimentEntity" extends="BaseEntityResultMap">
-    <result column="experiment_spec" jdbcType="VARCHAR" property="experimentSpec" />
-    <result column="experiment_status" property="experimentStatus"/>
-    <result column="accepted_time" property="acceptedTime"/>
-    <result column="running_time" property="runningTime"/>
-    <result column="finished_time" property="finishedTime"/>
-    <result column="uid" property="uid"/>
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, experiment_spec, create_by, create_time, update_by, update_time, experiment_status, accepted_time, running_time, finished_time, uid
-  </sql>
-
-  <select id="selectAll" parameterType="java.lang.String" resultMap="ExperimentEntityResultMap">
-    select
-    <include refid="Base_Column_List" />
-    from experiment
-  </select>
-
-  <select id="select" parameterType="java.lang.String" resultMap="ExperimentEntityResultMap">
-    select
-    <include refid="Base_Column_List" />
-    from experiment
-    where id = #{id,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="delete" parameterType="java.lang.String">
-    delete from experiment
-    where id = #{id,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.experiment.database.entity.ExperimentEntity">
-    insert into experiment (id, experiment_spec, create_by, create_time, update_by, update_time, experiment_status, accepted_time, running_time, finished_time, uid)
-    values (#{id,jdbcType=VARCHAR}, #{experimentSpec,jdbcType=VARCHAR},
-            #{createBy,jdbcType=VARCHAR}, now(), #{updateBy,jdbcType=VARCHAR}, now(), #{experimentStatus,jdbcType=VARCHAR},
-            #{acceptedTime,jdbcType=TIMESTAMP}, #{runningTime,jdbcType=TIMESTAMP}, #{finishedTime,jdbcType=TIMESTAMP}, #{uid,jdbcType=VARCHAR})
-  </insert>
-
-  <update id="update" parameterType="org.apache.submarine.server.experiment.database.entity.ExperimentEntity">
-    update experiment
-    <set>
-      <if test="experimentSpec != null and experimentStatus != null">
-        create_time = #{createTime, jdbcType=TIMESTAMP},
-        experiment_spec = #{experimentSpec,jdbcType=VARCHAR},
-        experiment_status = #{experimentStatus, jdbcType=VARCHAR},
-        accepted_time = #{acceptedTime,jdbcType=TIMESTAMP},
-        running_time = #{runningTime,jdbcType=TIMESTAMP},
-        finished_time = #{finishedTime,jdbcType=TIMESTAMP},
-        uid = #{uid,jdbcType=VARCHAR},
-      </if>
-      update_time = now()
-    </set>
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-</mapper>
