--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/ModelVersionTagMapper.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.model.mappers.ModelVersionTagMapper">
-  <resultMap id="resultMap" type="org.apache.submarine.server.model.database.entities.ModelVersionTagEntity">
-    <result column="name" property="name" />
-    <result column="tag" property="tag" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    name, tag
-  </sql>
-
-  <delete id="delete" parameterType="java.lang.String">
-    delete from model_version_tag
-    where name = #{name,jdbcType=VARCHAR} AND tag = #{tag,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.model.database.entities.ModelVersionTagEntity">
-    insert into model_version_tag (name, version, tag)
-    values (#{name,jdbcType=VARCHAR}, #{version,jdbcType=INTEGER}, #{tag,jdbcType=VARCHAR})
-  </insert>
-
-</mapper>
