--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/NotebookMapper.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.notebook.mappers.NotebookMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="NotebookEntityResultMap" type="org.apache.submarine.server.notebook.database.entity.NotebookEntity" extends="BaseEntityResultMap">
-    <result column="notebook_spec" jdbcType="VARCHAR" property="notebookSpec" />
-    <result column="notebook_status" jdbcType="VARCHAR" property="notebookStatus" />
-    <result column="notebook_url" jdbcType="VARCHAR" property="notebookUrl" />
-    <result column="reason" jdbcType="VARCHAR" property="reason" />
-    <result column="deleted_time" property="deletedTime" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, notebook_spec, create_by, create_time, update_by, update_time, notebook_status, notebook_url, reason, deleted_time
-  </sql>
-
-  <select id="selectAll" parameterType="java.lang.String" resultMap="NotebookEntityResultMap">
-    select
-    <include refid="Base_Column_List" />
-    from notebook
-  </select>
-
-  <select id="select" parameterType="java.lang.String" resultMap="NotebookEntityResultMap">
-    select
-    <include refid="Base_Column_List" />
-    from notebook
-    where id = #{id,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="delete" parameterType="java.lang.String">
-    delete from notebook
-    where id = #{id,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.notebook.database.entity.NotebookEntity">
-    insert into notebook (id, notebook_spec, create_by, create_time, update_by, update_time,
-     notebook_status, notebook_url, reason, deleted_time)
-    values (#{id,jdbcType=VARCHAR}, #{notebookSpec,jdbcType=VARCHAR},
-            #{createBy,jdbcType=VARCHAR}, #{createTime,jdbcType=TIMESTAMP}, #{updateBy,jdbcType=VARCHAR},
-            now(), #{notebookStatus,jdbcType=VARCHAR}, #{notebookUrl,jdbcType=VARCHAR},
-            #{reason,jdbcType=VARCHAR},null)
-  </insert>
-
-  <update id="update" parameterType="org.apache.submarine.server.notebook.database.entity.NotebookEntity">
-    update notebook
-    <set>
-      <if test="notebookSpec != null and notebookStatus != null">
-        notebook_spec = #{notebookSpec,jdbcType=VARCHAR},
-        create_time = #{createTime,jdbcType=TIMESTAMP},
-        notebook_status = #{notebookStatus,jdbcType=VARCHAR},
-        notebook_url = #{notebookUrl,jdbcType=VARCHAR},
-        reason = #{reason,jdbcType=VARCHAR},
-        deleted_time = #{deletedTime,jdbcType=TIMESTAMP},
-      </if>
-      update_time = now()
-    </set>
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-</mapper>
