--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/ProjectFilesMapper.xml
@@ -1,170 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.ProjectFilesMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity" extends="BaseEntityResultMap">
-    <result column="project_id" jdbcType="VARCHAR" property="projectId" />
-    <result column="file_name" jdbcType="VARCHAR" property="fileName" />
-    <result column="file_content" jdbcType="LONGVARCHAR" property="fileContent" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, project_id, file_name, file_content, create_by, create_time, update_by, update_time
-  </sql>
-
-  <select id="selectAll" parameterType="java.util.Map" resultMap="resultMap">
-    SELECT
-    <include refid="Base_Column_List"/>
-    FROM project_files
-    WHERE 1 = 1
-    <if test="projectId != null">
-      AND project_id = #{projectId,jdbcType=VARCHAR}
-    </if>
-  </select>
-
-  <select id="selectByPrimaryKey" parameterType="java.lang.String" resultMap="resultMap">
-    select
-    <include refid="Base_Column_List" />
-    from project_files
-    where id = #{id,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="deleteByPrimaryKey" parameterType="java.lang.String">
-    delete from project_files
-    where id = #{id,jdbcType=VARCHAR}
-  </delete>
-
-  <delete id="deleteSelective" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    delete from project_files
-    where 1 = 1
-    <if test="projectId != null">
-      AND project_id = #{projectId,jdbcType=VARCHAR}
-    </if>
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    <selectKey keyProperty="id" resultType="java.lang.String" order="BEFORE">
-      SELECT REPLACE(UUID(),"-","")
-    </selectKey>
-    insert into project_files (id, project_id, file_name,
-      create_by, create_time, update_by, update_time, file_content)
-    values (#{id,jdbcType=VARCHAR}, #{projectId,jdbcType=VARCHAR}, #{fileName,jdbcType=VARCHAR},
-      #{createBy,jdbcType=VARCHAR}, now(), #{updateBy,jdbcType=VARCHAR}, now(), #{fileContent,jdbcType=LONGVARCHAR})
-  </insert>
-
-  <insert id="insertSelective" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    insert into project_files
-    <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
-      <if test="projectId != null">
-        project_id,
-      </if>
-      <if test="fileName != null">
-        file_name,
-      </if>
-      <if test="createBy != null">
-        create_by,
-      </if>
-      <if test="updateBy != null">
-        update_by,
-      </if>
-      create_time, update_time,
-      <if test="fileContent != null">
-        file_content,
-      </if>
-    </trim>
-    <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=VARCHAR},
-      </if>
-      <if test="projectId != null">
-        #{projectId,jdbcType=VARCHAR},
-      </if>
-      <if test="fileName != null">
-        #{fileName,jdbcType=VARCHAR},
-      </if>
-      <if test="createBy != null">
-        #{createBy,jdbcType=VARCHAR},
-      </if>
-      <if test="updateBy != null">
-        #{updateBy,jdbcType=VARCHAR},
-      </if>
-      now(), now(),
-      <if test="fileContent != null">
-        #{fileContent,jdbcType=LONGVARCHAR},
-      </if>
-    </trim>
-  </insert>
-
-  <update id="updateByPrimaryKeySelective" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    update project_files
-    <set>
-      <if test="projectId != null">
-        project_id = #{projectId,jdbcType=VARCHAR},
-      </if>
-      <if test="fileName != null">
-        file_name = #{fileName,jdbcType=VARCHAR},
-      </if>
-      <if test="updateBy != null">
-        update_by = #{updateBy,jdbcType=VARCHAR},
-      </if>
-      <if test="updateTime != null">
-        update_time = #{updateTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="fileContent != null">
-        file_content = #{fileContent,jdbcType=LONGVARCHAR},
-      </if>
-    </set>
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-  <update id="updateByPrimaryKeyWithBLOBs" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    update project_files
-    set project_id = #{projectId,jdbcType=VARCHAR},
-      file_name = #{fileName,jdbcType=VARCHAR},
-      create_by = #{createBy,jdbcType=VARCHAR},
-      create_time = #{createTime,jdbcType=TIMESTAMP},
-      update_by = #{updateBy,jdbcType=VARCHAR},
-      update_time = #{updateTime,jdbcType=TIMESTAMP},
-      file_content = #{fileContent,jdbcType=LONGVARCHAR}
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-  <update id="updateByPrimaryKey" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectFilesEntity">
-    update project_files
-    set project_id = #{projectId,jdbcType=VARCHAR},
-      file_name = #{fileName,jdbcType=VARCHAR},
-      create_by = #{createBy,jdbcType=VARCHAR},
-      create_time = #{createTime,jdbcType=TIMESTAMP},
-      update_by = #{updateBy,jdbcType=VARCHAR},
-      update_time = #{updateTime,jdbcType=TIMESTAMP}
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-</mapper>
