--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/ProjectMapper.xml
@@ -1,205 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.ProjectMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.ProjectEntity" extends="BaseEntityResultMap">
-    <result column="name" jdbcType="VARCHAR" property="name" />
-    <result column="visibility" jdbcType="VARCHAR" property="visibility" />
-    <result column="permission" jdbcType="VARCHAR" property="permission" />
-    <result column="type" jdbcType="VARCHAR" property="type" />
-    <result column="description" jdbcType="VARCHAR" property="description" />
-    <result column="user_name" jdbcType="VARCHAR" property="userName" />
-    <result column="team_name" jdbcType="VARCHAR" property="teamName" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, name, visibility, permission, type, description, user_name, team_name,
-    create_by, create_time, update_by, update_time
-  </sql>
-
-  <select id="selectAll" parameterType="java.util.Map" resultMap="resultMap">
-    SELECT
-    <include refid="Base_Column_List"/>
-    FROM project
-    WHERE 1 = 1
-    <if test="userName!=null and userName!=''">AND `user_name` = #{userName}</if>
-    ORDER BY #{column} #{order}
-  </select>
-
-  <select id="selectByPrimaryKey" parameterType="java.lang.String" resultMap="resultMap">
-    select
-    <include refid="Base_Column_List" />
-    from project
-    where id = #{id,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="deleteByPrimaryKey" parameterType="java.lang.String">
-    delete from project
-    where id = #{id,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectEntity">
-    <selectKey keyProperty="id" resultType="java.lang.String" order="BEFORE">
-      SELECT REPLACE(UUID(),"-","")
-    </selectKey>
-    insert into project (id, name, visibility,
-      permission, type, description,
-      user_name, team_name,
-      create_by, create_time, update_by, update_time)
-    values (#{id,jdbcType=VARCHAR}, #{name,jdbcType=VARCHAR}, #{visibility,jdbcType=VARCHAR},
-      #{permission,jdbcType=VARCHAR}, #{type,jdbcType=VARCHAR}, #{description,jdbcType=VARCHAR},
-      #{userName,jdbcType=VARCHAR}, #{teamName,jdbcType=VARCHAR},
-      #{createBy,jdbcType=VARCHAR}, now(), #{updateBy,jdbcType=VARCHAR}, now())
-  </insert>
-
-  <insert id="insertSelective" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectEntity">
-    insert into project
-    <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
-      <if test="name != null">
-        name,
-      </if>
-      <if test="visibility != null">
-        visibility,
-      </if>
-      <if test="permission != null">
-        permission,
-      </if>
-      <if test="type != null">
-        type,
-      </if>
-      <if test="description != null">
-        description,
-      </if>
-      <if test="userName != null">
-        user_name,
-      </if>
-      <if test="teamName != null">
-        team_name,
-      </if>
-      <if test="createBy != null">
-        create_by,
-      </if>
-      <if test="createTime != null">
-        create_time,
-      </if>
-      <if test="updateBy != null">
-        update_by,
-      </if>
-      <if test="updateTime != null">
-        update_time,
-      </if>
-    </trim>
-    <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=VARCHAR},
-      </if>
-      <if test="name != null">
-        #{name,jdbcType=VARCHAR},
-      </if>
-      <if test="visibility != null">
-        #{visibility,jdbcType=VARCHAR},
-      </if>
-      <if test="permission != null">
-        #{permission,jdbcType=VARCHAR},
-      </if>
-      <if test="type != null">
-        #{type,jdbcType=VARCHAR},
-      </if>
-      <if test="description != null">
-        #{description,jdbcType=VARCHAR},
-      </if>
-      <if test="userName != null">
-        #{userName,jdbcType=VARCHAR},
-      </if>
-      <if test="teamName != null">
-        #{teamName,jdbcType=VARCHAR},
-      </if>
-      <if test="createBy != null">
-        #{createBy,jdbcType=VARCHAR},
-      </if>
-      <if test="createTime != null">
-        #{createTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="updateBy != null">
-        #{updateBy,jdbcType=VARCHAR},
-      </if>
-      <if test="updateTime != null">
-        #{updateTime,jdbcType=TIMESTAMP},
-      </if>
-    </trim>
-  </insert>
-
-  <update id="updateByPrimaryKeySelective" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectEntity">
-    update project
-    <set>
-      <if test="name != null">
-        name = #{name,jdbcType=VARCHAR},
-      </if>
-      <if test="visibility != null">
-        visibility = #{visibility,jdbcType=VARCHAR},
-      </if>
-      <if test="permission != null">
-        permission = #{permission,jdbcType=VARCHAR},
-      </if>
-      <if test="type != null">
-        type = #{type,jdbcType=VARCHAR},
-      </if>
-      <if test="description != null">
-        description = #{description,jdbcType=VARCHAR},
-      </if>
-      <if test="userName != null">
-        user_name = #{userName,jdbcType=VARCHAR},
-      </if>
-      <if test="teamName != null">
-        team_name = #{teamName,jdbcType=VARCHAR},
-      </if>
-      <if test="updateBy != null">
-        update_by = #{updateBy,jdbcType=VARCHAR},
-      </if>
-      update_time = now()
-    </set>
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-  <update id="updateByPrimaryKey" parameterType="org.apache.submarine.server.database.workbench.entity.ProjectEntity">
-    update project
-    set name = #{name,jdbcType=VARCHAR},
-      visibility = #{visibility,jdbcType=VARCHAR},
-      permission = #{permission,jdbcType=VARCHAR},
-      type = #{type,jdbcType=VARCHAR},
-      description = #{description,jdbcType=VARCHAR},
-      user_name = #{userName,jdbcType=VARCHAR},
-      team_name = #{teamName,jdbcType=VARCHAR},
-      update_by = #{updateBy,jdbcType=VARCHAR},
-      update_time = #{updateTime,jdbcType=TIMESTAMP}
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-</mapper>
