--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/SysDeptMapper.xml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.SysDeptMapper">
-
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id" />
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.SysDeptEntity" extends="BaseEntityResultMap">
-    <result column="dept_code" property="deptCode"/>
-    <result column="dept_name" property="deptName"/>
-    <result column="parent_code" property="parentCode"/>
-    <result column="parent_name" property="parentName"/>
-    <result column="sort_order" property="sortOrder"/>
-    <result column="description" property="description"/>
-    <result column="deleted" property="deleted"/>
-  </resultMap>
-
-  <select id="selectAll" parameterType="java.util.Map" resultMap="resultMap">
-    SELECT a.*, b.dept_name AS parent_name
-    FROM sys_department a LEFT JOIN sys_department b ON a.parent_code=b.dept_code
-    WHERE 1=1
-    <if test="deptCode!=null and deptCode!=''"> AND a.`dept_code` like '%${deptCode}%' </if>
-    <if test="deptName!=null and deptName!=''"> AND a.`dept_name` like '%${deptName}%' </if>
-    ORDER BY a.sort_order
-  </select>
-
-  <insert id="add" parameterType="org.apache.submarine.server.database.workbench.entity.SysDeptEntity">
-    <selectKey keyProperty="id" resultType="String" order="BEFORE">
-      select REPLACE(UUID(),"-","")
-    </selectKey>
-    INSERT INTO sys_department (`id`, `dept_code`, `dept_name`,
-    <if test="parentCode!=null and parentCode!=''"> `parent_code`, </if>
-    <if test="deleted!=null"> `deleted`, </if>
-    `sort_order`, `description`,
-    `create_by`, `create_time`, `update_time`)
-    VALUES (
-    #{id}, #{deptCode}, #{deptName},
-    <if test="parentCode!=null and parentCode!=''"> #{parentCode}, </if>
-    <if test="deleted!=null">#{deleted}, </if>
-    #{sortOrder}, #{description},
-    #{createBy}, now(), now())
-  </insert>
-
-  <select id="getById" parameterType="java.lang.String" resultMap="resultMap">
-    SELECT * FROM sys_department WHERE id = #{id}
-  </select>
-
-  <delete id="deleteById" parameterType="java.lang.String">
-    DELETE FROM sys_department WHERE id = #{id}
-  </delete>
-
-  <update id="updateBy" parameterType="org.apache.submarine.server.database.workbench.entity.SysDeptEntity">
-    UPDATE sys_department SET
-    <if test="deptCode!=null and deptCode!=''">`dept_code` = #{deptCode},</if>
-    <if test="deptName!=null and deptName!=''">`dept_name` = #{deptName},</if>
-    <choose>
-      <when test="parentCode!=null and parentCode!=''">
-        `parent_code` = #{parentCode},
-      </when>
-      <otherwise>
-        `parent_code` = NULL,
-      </otherwise>
-    </choose>
-    <if test="sortOrder!=null">`sort_order` = #{sortOrder},</if>
-    <if test="description!=null and description!=''">`description` = #{description},</if>
-    <if test="deleted!=null">`deleted` = #{deleted},</if>
-    <if test="createBy!=null">`create_by` = #{createBy},</if>
-    <if test="createTime!=null">`create_time` = #{createTime},</if>
-    <if test="updateBy!=null">`update_by` = #{updateBy},</if>
-    `update_time`=now()
-    WHERE id = #{id}
-  </update>
-
-  <update id="deleteBatch">
-    UPDATE sys_department SET deleted = 1, `update_time` = now() WHERE id IN
-    <foreach collection="list" item="id" index="index" open="(" close=")" separator=",">
-      #{id}
-    </foreach>
-  </update>
-
-  <update id="resetDeptLevel">
-    UPDATE sys_department SET parent_code = NULL, `update_time` = now()
-  </update>
-</mapper>
