--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/SysDictItemMapper.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.SysDictItemMapper">
-
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id" />
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <select id="selectAll" resultMap="resultMap">
-    SELECT * FROM sys_dict_item WHERE 1 = 1
-    <if test="dictCode!=null and dictCode!=''"> AND `dict_code` = #{dictCode}</if>
-    <if test="itemCode!=null and itemCode!=''"> AND `item_code` like '%${itemCode}%'</if>
-    <if test="itemName!=null and itemName!=''"> AND `item_name` like '%${itemName}%'</if>
-    ORDER BY sort_order
-  </select>
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.SysDictItemEntity" extends="BaseEntityResultMap">
-    <result column="dict_code" property="dictCode"/>
-    <result column="item_code" property="itemCode"/>
-    <result column="item_name" property="itemName"/>
-    <result column="description" property="description"/>
-    <result column="sort_order" property="sortOrder"/>
-    <result column="deleted" property="deleted"/>
-  </resultMap>
-
-  <insert id="insertSysDictItem" parameterType="org.apache.submarine.server.database.workbench.entity.SysDictItemEntity">
-    INSERT INTO sys_dict_item (`id`, `dict_code`, `item_code`, `item_name`, `description`,
-    <if test="sortOrder!=null"> `sort_order`, </if>
-    <if test="deleted!=null"> `deleted`, </if>
-    `create_by`, `create_time`, `update_by`, `update_time`)
-    VALUES (REPLACE(UUID(), '-', ''), #{dictCode}, #{itemCode}, #{itemName}, #{description},
-    <if test="sortOrder!=null">#{sortOrder}, </if>
-    <if test="deleted!=null">#{deleted}, </if>
-    #{createBy}, now(), #{updateBy}, now())
-  </insert>
-
-  <select id="getById" parameterType="java.lang.String" resultMap="resultMap">
-    SELECT * FROM sys_dict_item WHERE id = #{id}
-  </select>
-
-  <delete id="deleteById" parameterType="java.lang.String">
-    DELETE FROM sys_dict_item WHERE id = #{id}
-  </delete>
-
-  <update id="updateBy" parameterType="org.apache.submarine.server.database.workbench.entity.SysDictItemEntity">
-    UPDATE sys_dict_item SET
-    <if test="dictCode!=null and dictCode!=''">`dict_code` = #{dictCode},</if>
-    <if test="itemCode!=null and itemCode!=''">`item_code` = #{itemCode},</if>
-    <if test="itemName!=null and itemName!=''">`item_name` = #{itemName},</if>
-    <if test="description!=null and description != ''">`description` = #{description},</if>
-    <if test="sortOrder!=null">`sort_order` = #{sortOrder},</if>
-    <if test="deleted!=null">`deleted` = #{deleted},</if>
-    <if test="createBy!=null">`create_by` = #{createBy},</if>
-    <if test="createTime!=null">`create_time` = #{createTime},</if>
-    <if test="updateBy!=null">`update_by` = #{updateBy},</if>
-    `update_time` = now()
-    WHERE id = #{id}
-  </update>
-
-  <select id="queryDictByCode" parameterType="java.util.Map" resultMap="resultMap">
-    SELECT `item_code`, `item_name` FROM sys_dict_item WHERE `dict_code` = #{dictCode}
-  </select>
-</mapper>
