--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/SysDictMapper.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.SysDictMapper">
-
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <select id="selectAll" parameterType="java.util.Map" resultMap="resultMap">
-    SELECT * FROM sys_dict
-    WHERE 1=1
-    <if test="dictCode!=null and dictCode!=''">AND `dict_code` like '%${dictCode}%'</if>
-    <if test="dictName!=null and dictName!=''">AND `dict_name` like '%${dictName}%'</if>
-    ORDER BY id
-  </select>
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.SysDictEntity" extends="BaseEntityResultMap">
-    <result column="dict_code" property="dictCode"/>
-    <result column="dict_name" property="dictName"/>
-    <result column="description" property="description"/>
-    <result column="deleted" property="deleted"/>
-    <result column="type" property="type"/>
-  </resultMap>
-
-  <insert id="insertSysDict" parameterType="org.apache.submarine.server.database.workbench.entity.SysDictEntity">
-    INSERT INTO sys_dict(`id`, `dict_code`, `dict_name`, `description`,
-    <if test="deleted!=null">`deleted`,</if>
-    <if test="type!=null">`type`,</if>
-    `create_by`, `create_time`, `update_time`)
-    VALUES (
-    <choose>
-      <when test="id!=null and id!=''">
-        #{id},
-      </when>
-      <otherwise>
-        REPLACE(UUID(), '-', ''),
-      </otherwise>
-    </choose>
-    #{dictCode}, #{dictName}, #{description},
-    <if test="deleted!=null">#{deleted},</if>
-    <if test="type!=null">#{type},</if>
-    #{createBy}, now(), now())
-  </insert>
-
-  <select id="getById" parameterType="java.lang.String" resultMap="resultMap">
-    SELECT * FROM sys_dict WHERE id = #{id}
-  </select>
-
-  <delete id="deleteById" parameterType="java.lang.String">
-    DELETE FROM sys_dict WHERE id = #{id}
-  </delete>
-
-  <update id="updateBy" parameterType="org.apache.submarine.server.database.workbench.entity.SysDictEntity">
-    UPDATE sys_dict SET
-    <if test="dictCode!=null and dictCode!=''">`dict_code`=#{dictCode},</if>
-    <if test="dictName!=null and dictName!=''">`dict_name`=#{dictName},</if>
-    <if test="description!=null and description!=''">`description`=#{description},</if>
-    <if test="deleted!=null">`deleted`=#{deleted},</if>
-    <if test="type!=null">`type`=#{type},</if>
-    <if test="createBy!=null">`create_by`=#{createBy},</if>
-    <if test="createTime!=null">`create_time`=#{createTime},</if>
-    <if test="updateBy!=null">`update_by`=#{updateBy},</if>
-    `update_time`=now()
-    WHERE id = #{id}
-  </update>
-</mapper>
