--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/SysMessageMapper.xml
@@ -1,202 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.SysMessageMapper">
-  <resultMap id="BaseEntityResultMap" type="org.apache.submarine.server.database.entity.BaseEntity">
-    <id property="id" column="id"/>
-    <result column="create_by" property="createBy"/>
-    <result column="create_time" property="createTime"/>
-    <result column="update_by" property="updateBy"/>
-    <result column="update_time" property="updateTime"/>
-  </resultMap>
-
-  <resultMap id="resultMap" type="org.apache.submarine.server.database.workbench.entity.SysMessageEntity" extends="BaseEntityResultMap">
-    <result column="sender" jdbcType="VARCHAR" property="sender"/>
-    <result column="receiver" jdbcType="VARCHAR" property="receiver"/>
-    <result column="type" jdbcType="VARCHAR" property="type"/>
-    <result column="status" jdbcType="INTEGER" property="status"/>
-    <result column="context" jdbcType="LONGVARCHAR" property="context"/>
-  </resultMap>
-
-  <resultMap extends="resultMap" id="ResultMapWithBLOBs" type="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    <result column="context" jdbcType="LONGVARCHAR" property="context"/>
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    id, sender, receiver, type, status, create_by, create_time, update_by, update_time
-  </sql>
-  <sql id="Blob_Column_List">
-    context
-  </sql>
-
-  <select id="selectByPrimaryKey" parameterType="java.lang.String" resultMap="ResultMapWithBLOBs">
-    select
-    <include refid="Base_Column_List"/>
-    ,
-    <include refid="Blob_Column_List"/>
-    from sys_message
-    where id = #{id,jdbcType=VARCHAR}
-  </select>
-
-  <delete id="deleteByPrimaryKey" parameterType="java.lang.String">
-    delete from sys_message
-    where id = #{id,jdbcType=VARCHAR}
-  </delete>
-
-  <insert id="insert" parameterType="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    insert into sys_message (id, sender, receiver,
-      type, status, create_by,
-      create_time, update_time,
-      context)
-    values (REPLACE(UUID(),"-",""), #{sender,jdbcType=VARCHAR}, #{receiver,jdbcType=VARCHAR},
-      #{type,jdbcType=VARCHAR}, #{status,jdbcType=INTEGER}, #{createBy,jdbcType=VARCHAR},
-      now(), now()
-      #{context,jdbcType=LONGVARCHAR})
-  </insert>
-
-  <insert id="insertSelective" parameterType="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    insert into sys_message
-    <trim prefix="(" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        id,
-      </if>
-      <if test="sender != null">
-        sender,
-      </if>
-      <if test="receiver != null">
-        receiver,
-      </if>
-      <if test="type != null">
-        type,
-      </if>
-      <if test="status != null">
-        status,
-      </if>
-      <if test="createBy != null">
-        create_by,
-      </if>
-      <if test="createTime != null">
-        create_time,
-      </if>
-      <if test="updateBy != null">
-        update_by,
-      </if>
-      <if test="updateTime != null">
-        update_time,
-      </if>
-      <if test="context != null">
-        context,
-      </if>
-    </trim>
-    <trim prefix="values (" suffix=")" suffixOverrides=",">
-      <if test="id != null">
-        #{id,jdbcType=VARCHAR},
-      </if>
-      <if test="sender != null">
-        #{sender,jdbcType=VARCHAR},
-      </if>
-      <if test="receiver != null">
-        #{receiver,jdbcType=VARCHAR},
-      </if>
-      <if test="type != null">
-        #{type,jdbcType=VARCHAR},
-      </if>
-      <if test="status != null">
-        #{status,jdbcType=INTEGER},
-      </if>
-      <if test="createBy != null">
-        #{createBy,jdbcType=VARCHAR},
-      </if>
-      <if test="createTime != null">
-        #{createTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="updateBy != null">
-        #{updateBy,jdbcType=VARCHAR},
-      </if>
-      <if test="updateTime != null">
-        #{updateTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="context != null">
-        #{context,jdbcType=LONGVARCHAR},
-      </if>
-    </trim>
-  </insert>
-
-  <update id="updateByPrimaryKeySelective" parameterType="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    update sys_message
-    <set>
-      <if test="sender != null">
-        sender = #{sender,jdbcType=VARCHAR},
-      </if>
-      <if test="receiver != null">
-        receiver = #{receiver,jdbcType=VARCHAR},
-      </if>
-      <if test="type != null">
-        type = #{type,jdbcType=VARCHAR},
-      </if>
-      <if test="status != null">
-        status = #{status,jdbcType=INTEGER},
-      </if>
-      <if test="createBy != null">
-        create_by = #{createBy,jdbcType=VARCHAR},
-      </if>
-      <if test="createTime != null">
-        create_time = #{createTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="updateBy != null">
-        update_by = #{updateBy,jdbcType=VARCHAR},
-      </if>
-      <if test="updateTime != null">
-        update_time = #{updateTime,jdbcType=TIMESTAMP},
-      </if>
-      <if test="context != null">
-        context = #{context,jdbcType=LONGVARCHAR},
-      </if>
-    </set>
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-  <update id="updateByPrimaryKeyWithBLOBs" parameterType="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    update sys_message
-    set sender = #{sender,jdbcType=VARCHAR},
-      receiver = #{receiver,jdbcType=VARCHAR},
-      type = #{type,jdbcType=VARCHAR},
-      status = #{status,jdbcType=INTEGER},
-      create_by = #{createBy,jdbcType=VARCHAR},
-      create_time = #{createTime,jdbcType=TIMESTAMP},
-      update_by = #{updateBy,jdbcType=VARCHAR},
-      update_time = #{updateTime,jdbcType=TIMESTAMP},
-      context = #{context,jdbcType=LONGVARCHAR}
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-
-  <update id="updateByPrimaryKey" parameterType="org.apache.submarine.server.database.workbench.entity.SysMessageEntity">
-    update sys_message
-    set sender = #{sender,jdbcType=VARCHAR},
-      receiver = #{receiver,jdbcType=VARCHAR},
-      type = #{type,jdbcType=VARCHAR},
-      status = #{status,jdbcType=INTEGER},
-      create_by = #{createBy,jdbcType=VARCHAR},
-      create_time = #{createTime,jdbcType=TIMESTAMP},
-      update_by = #{updateBy,jdbcType=VARCHAR},
-      update_time = #{updateTime,jdbcType=TIMESTAMP}
-    where id = #{id,jdbcType=VARCHAR}
-  </update>
-</mapper>
