--- submarine-server/server-core/src/main/resources/org/apache/submarine/database/mappers/SystemMapper.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.submarine.server.database.workbench.mappers.SystemMapper">
-
-  <select id="duplicateCheck" resultType="Long"  parameterType="java.util.Map">
-    SELECT COUNT(*) FROM ${tableName} WHERE ${fieldName} = #{fieldVal}
-    <if test="equalFieldName!=null and equalFieldName!='' and equalFieldVal!=null and equalFieldVal!=''">
-      and ${equalFieldName} = #{equalFieldVal}
-    </if>
-    <if test="dataId!=null and dataId!=''">
-      and id &lt;&gt; #{dataId}
-    </if>
-  </select>
-</mapper>
