+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/model/database/ModelVersionTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.submarine.server.model.database;
+package org.apache.submarine.server.database.model;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -25,10 +25,10 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.submarine.server.model.database.entities.ModelVersionEntity;
-import org.apache.submarine.server.model.database.entities.RegisteredModelEntity;
-import org.apache.submarine.server.model.database.service.ModelVersionService;
-import org.apache.submarine.server.model.database.service.RegisteredModelService;
+import org.apache.submarine.server.database.model.entities.ModelVersionEntity;
+import org.apache.submarine.server.database.model.entities.RegisteredModelEntity;
+import org.apache.submarine.server.database.model.service.ModelVersionService;
+import org.apache.submarine.server.database.model.service.RegisteredModelService;
 
 public class ModelVersionTest {
   RegisteredModelService registeredModelService = new RegisteredModelService();
