+++ submarine-server/server-database/src/main/java/org/apache/submarine/server/database/experiment/ExperimentManager.java
@@ -42,8 +42,8 @@ import org.apache.submarine.server.api.experiment.ExperimentLog;
 import org.apache.submarine.server.api.experiment.TensorboardInfo;
 import org.apache.submarine.server.api.experiment.MlflowInfo;
 import org.apache.submarine.server.api.spec.ExperimentSpec;
-import org.apache.submarine.server.experiment.database.entity.ExperimentEntity;
-import org.apache.submarine.server.experiment.database.service.ExperimentService;
+import org.apache.submarine.server.database.experiment.entity.ExperimentEntity;
+import org.apache.submarine.server.database.experiment.service.ExperimentService;
 import org.apache.submarine.server.rest.RestConstants;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
