+++ submarine-server/server-database/src/main/java/org/apache/submarine/server/database/experiment/service/ExperimentService.java
@@ -22,7 +22,7 @@ package org.apache.submarine.server.database.experiment.service;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.submarine.commons.utils.exception.SubmarineRuntimeException;
 import org.apache.submarine.server.database.database.utils.MyBatisUtil;
-import org.apache.submarine.server.experiment.database.entity.ExperimentEntity;
+import org.apache.submarine.server.database.experiment.entity.ExperimentEntity;
 import org.apache.submarine.server.database.experiment.mappers.ExperimentMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
