+++ submarine-server/server-database/src/main/java/org/apache/submarine/server/database/model/service/ModelVersionService.java
@@ -22,7 +22,7 @@ package org.apache.submarine.server.database.model.service;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.submarine.commons.utils.exception.SubmarineRuntimeException;
 import org.apache.submarine.server.database.database.utils.MyBatisUtil;
-import org.apache.submarine.server.model.database.entities.ModelVersionEntity;
+import org.apache.submarine.server.database.model.entities.ModelVersionEntity;
 import org.apache.submarine.server.database.model.mappers.ModelVersionMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
