+++ submarine-server/server-database/src/main/java/org/apache/submarine/server/database/model/service/RegisteredModelTagService.java
@@ -22,7 +22,7 @@ package org.apache.submarine.server.database.model.service;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.submarine.commons.utils.exception.SubmarineRuntimeException;
 import org.apache.submarine.server.database.database.utils.MyBatisUtil;
-import org.apache.submarine.server.model.database.entities.RegisteredModelTagEntity;
+import org.apache.submarine.server.database.model.entities.RegisteredModelTagEntity;
 import org.apache.submarine.server.database.model.mappers.RegisteredModelTagMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
