+++ submarine-server/server-database/src/main/java/org/apache/submarine/server/database/notebook/NotebookManager.java
@@ -35,7 +35,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
-import org.apache.submarine.server.notebook.database.service.NotebookService;
+import org.apache.submarine.server.database.notebook.service.NotebookService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
