+++ submarine-server/server-database/src/main/resources/org/apache/submarine/database/mappers/ExperimentTemplateMapper.xml
@@ -27,7 +27,7 @@
     <result column="update_time" property="updateTime"/>
   </resultMap>
 
-  <resultMap id="resultMap" type="org.apache.submarine.server.experimenttemplate.database.entity.ExperimentTemplateEntity" extends="BaseEntityResultMap">
+  <resultMap id="resultMap" type="org.apache.submarine.server.database.experimenttemplate.entity.ExperimentTemplateEntity" extends="BaseEntityResultMap">
     <result column="experimentTemplate_name" jdbcType="VARCHAR" property="experimentTemplateName" />
     <result column="experimentTemplate_spec" jdbcType="VARCHAR" property="experimentTemplateSpec" />
   </resultMap>
@@ -48,7 +48,7 @@
     where experimentTemplate_name = #{experimentTemplate_name,jdbcType=VARCHAR}
   </delete>
 
-  <insert id="insert" parameterType="org.apache.submarine.server.experimenttemplate.database.entity.ExperimentTemplateEntity">
+  <insert id="insert" parameterType="org.apache.submarine.server.database.experimenttemplate.entity.ExperimentTemplateEntity">
     insert into experiment_template (id, experimentTemplate_name, experimentTemplate_spec,
       create_by, create_time, update_by, update_time)
     values (#{id,jdbcType=VARCHAR}, 
@@ -57,7 +57,7 @@
     #{createBy,jdbcType=VARCHAR}, now(), #{updateBy,jdbcType=VARCHAR}, now())
   </insert>
 
-  <update id="update" parameterType="org.apache.submarine.server.experimenttemplate.database.entity.ExperimentTemplateEntity">
+  <update id="update" parameterType="org.apache.submarine.server.database.experimenttemplate.entity.ExperimentTemplateEntity">
     update experiment_template
     <set>
       <if test="experimentTemplateSpec != null">
